rating

+++ resources/js/users.js
@@ -8,9 +8,6 @@ $(document).ready(function() {
     });
 
     $(document).on('click', '.buttonRole', function() {
-        const thisRef = $(this);
-        thisRef.text('processing');
-
         $.ajax({
             type: 'POST',
             url: 'ajaxChangeRole',
@@ -19,7 +16,6 @@ $(document).ready(function() {
                 userID: $('.userID').val()
             },
             success:function(response) {
-                console.log(response);
                 list();
                 showAlert(200, 'Role changed successfully');
 
@@ -47,17 +43,20 @@ function list() {
                                 <th>Action</th>
                             </tr>
                             </thead>
-            <tbody>
-        </tbody>
-    </table>`);
+                            <tbody>
+                            </tbody>
+                            </table>`);
 
             response.forEach(element => {
                 $('.userList tbody').append(`<tr>
                         <th scope="row" class="counterCell"></th>
                         <td>${element.name}</td>
                         <td>
+                        ` + ((user.id == element.id || user.role == 'admin') ? `
                         <a href="user/${element.id}/edit" title="Edit" class="btn btn-sm btn-primary">Edit</a>
-                        <a href="user/${element.id}/delete" title="Delete" data-method="DELETE" class="btn btn-sm btn-danger" data-confrim="Are you sure?">Delete</a>
+                        <a href="user/${element.id}/delete" title="Delete" data-method="DELETE" class="btn btn-sm btn-danger" onclick="return confirm('Are you sure you want to delete user ${element.name}');" data-confrim="Are you sure?">Delete</a>
+                        ` : `` ) + `
+                        ` + ((user.role == 'admin') ? `
                         <div class="btn-group">
                         <button class="btn btn-sm btn-secondary dropdown-toggle userID" name="valueButton"  type="button" value="" id="dropdownMenuButton" onclick="setUserID(${element.id})" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                         ${element.role}
@@ -78,6 +77,7 @@ function list() {
                         }
                         </script>
                         </div>
+                        ` : `` ) + `
                         </td>
                 </tr>`);
             });
